Refactor handlers
Relates-to: BLAIS5-2892
